Add sceneDelegateBox support
